wip
git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@3417 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ lib/qffitalgorithm.cpp
@@ -612,7 +612,7 @@ void QFFitAlgorithm::IRLSFunctorAdaptor::irlsReweight(const double *params)
 {
     double* err=(double*)malloc(irls_functor->get_evalout()*sizeof(double));
 
-    if (irls_functor->get_evalout()>=irls_functor->getDataPoints()) {
+    if (irls_functor->get_evalout()>=(int64_t)irls_functor->getDataPoints()) {
         // calculate error vector (f(xi,p)-yi)/sigmai
         irls_functor->evaluate(err, params);
         for (int i=0; i<irls_functor->get_evalout(); i++) {
